backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
paddleocr==2.7.3
paddlepaddle==2.5.2
jieba==0.42.1
pypinyin==0.49.0
pydantic==2.5.0
transformers==4.35.0
torch==2.1.0
sentencepiece==0.1.99
Pillow==10.1.0
numpy==1.24.3
requests==2.31.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
